refactor: move update settings action and reorder collection action to dedicated route files
