+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.debug.reorder.tsx
@@ -1,0 +1,46 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import * as models from '../../models';
+import { getById, update } from '../../models/helpers/request-operations';
+import { isRequestGroup, isRequestGroupId } from '../../models/request-group';
+import { invariant } from '../../utils/invariant';
+
+const getCollectionItem = async (id: string) => {
+  let item;
+  if (isRequestGroupId(id)) {
+    item = await models.requestGroup.getById(id);
+  } else {
+    item = await getById(id);
+  }
+
+  invariant(item, 'Item not found');
+
+  return item;
+};
+
+export async function action({ request, params }: ActionFunctionArgs) {
+  const { workspaceId } = params;
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+  const { id, targetId, dropPosition, metaSortKey } = await request.json();
+  invariant(typeof id === 'string', 'ID is required');
+  invariant(typeof targetId === 'string', 'Target ID is required');
+  invariant(typeof dropPosition === 'string', 'Drop position is required');
+  invariant(typeof metaSortKey === 'number', 'MetaSortKey position is required');
+
+  if (id === targetId) {
+    return null;
+  }
+
+  const item = await getCollectionItem(id);
+  const targetItem = await getCollectionItem(targetId);
+
+  const parentId = dropPosition === 'after' && isRequestGroup(targetItem) ? targetItem._id : targetItem.parentId;
+
+  if (isRequestGroup(item)) {
+    await models.requestGroup.update(item, { parentId, metaSortKey });
+  } else {
+    await update(item, { parentId, metaSortKey });
+  }
+
+  return null;
+}
