+++ packages/insomnia/src/ui/routes/settings.update.tsx
@@ -1,0 +1,13 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import * as models from '../../models';
+import { SegmentEvent } from '../analytics';
+
+export async function action({ request }: ActionFunctionArgs) {
+  const patch = await request.json();
+  if ('enableAnalytics' in patch && !patch.enableAnalytics) {
+    window.main.trackSegmentEvent({ event: SegmentEvent.analyticsDisabled });
+  }
+  await models.settings.patch(patch);
+  return null;
+}
